+++ app/bot/action/Private/UserClearLoginAction.go
@@ -5,7 +5,7 @@ import (
 	"main.go/app/bot/model/UserTokenModel"
 )
 
-func App_userClearLogin(bot int, uid, gid int) {
+func App_userClearLogin(bot int64, uid, gid int64) {
 	if UserTokenModel.Api_delete(uid) {
 		if gid != 0 {
 			api.Sendgrouptempmsg(bot, gid, uid, "您的登录状态已经全部清空，如需再次登录请发送acfur登录", false)
