+++ app/bot/action/Private/bindgroup.go
@@ -5,7 +5,7 @@ import (
 	"main.go/app/bot/model/BotGroupAllowModel"
 )
 
-func App_bind_group(bot int, uid int, text string) {
+func App_bind_group(bot int64, uid int64, text string) {
 	if len(text) < 2 {
 		api.Sendprivatemsg(bot, uid, "请使用\"acfur绑定群群号\"，来绑定新群", false)
 		return
@@ -21,7 +21,7 @@ func App_bind_group(bot int, uid int, text string) {
 	}
 }
 
-func App_unbind_group(bot int, uid int, text string) {
+func App_unbind_group(bot int64, uid int64, text string) {
 	if len(text) < 2 {
 		api.Sendprivatemsg(bot, uid, "请使用\"acfur绑定群群号\"，来绑定新群", false)
 		return
